feat: Give player Collector's Map if grubs are not randomized.
Talking to Elder Bug 3 times, in addition to giving the maps and the
quill, also gives the Collector's Map (which shows the locations of all
grubs) if grubs are not randomized. If grubs ARE randomized, the
Collector's Map is more hindrance than help because the grub pins will
never be removed, as the grubs cannot actually be collected at the
indicated locations.

Manually tested the following test cases:
* Start a new game with default randomizer settings, talked to elderbug
  3 times, confirmed that grub pins show up.
* Start a new game with all randomizer settings enabled, talked to
  elderbug 3 times, confirmed that grub pins do not show up.